user id to string (/me returns type string), User models, forms and struct adjusted
