+++ Cargo.toml
@@ -31,6 +31,8 @@ serde_derive = "1.0.188"
 actix-web-httpauth = "0.8.1"
 actix-cors = "0.6.4"
 tracing-actix-web = "0.7.7"
+regex = "1.10.2"
+
 
 [dependencies.sqlx]
 version = "0.6.3"
