+++ src/forms/mod.rs
@@ -1,6 +1,7 @@
 mod rating;
 
 pub(crate) mod stack;
+pub mod user;
 
 pub use rating::*;
