+++ src/models/stack.rs
@@ -9,7 +9,7 @@ use uuid::Uuid;
 pub struct Stack {
     pub id: i32,       // id - is a unique identifier for the app stack
     pub stack_id: Uuid, // external stack ID
-    pub user_id: i32,  // external unique identifier for the user
+    pub user_id: String,  // external unique identifier for the user
     pub name: String,
     // pub body: sqlx::types::Json<String>,
     pub body: Value, //json type
